# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ДомПульс — API прототипа",
    "description": "Ядро обращений и транспорт webhook MAX. Внешние системы УК ещё не подключены.",
    "version": "0.1.0"
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/webhooks/max": {
      "post": {
        "summary": "Max Webhook",
        "operationId": "max_webhook_webhooks_max_post",
        "parameters": [
          {
            "name": "X-Max-Bot-Api-Secret",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Max-Bot-Api-Secret"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/me": {
      "get": {
        "summary": "Me",
        "operationId": "me_api_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Profile"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/metrics/house": {
      "get": {
        "summary": "Metrics",
        "operationId": "metrics_api_metrics_house_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/tickets": {
      "post": {
        "summary": "Create Ticket",
        "operationId": "create_ticket_api_tickets_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TicketCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List Tickets",
        "operationId": "list_tickets_api_tickets_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Ticket"
                  },
                  "title": "Response List Tickets Api Tickets Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tickets/{ticket_id}": {
      "get": {
        "summary": "Get Ticket",
        "operationId": "get_ticket_api_tickets__ticket_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "ticket_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticket Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tickets/{ticket_id}/comments": {
      "post": {
        "summary": "Add Comment",
        "operationId": "add_comment_api_tickets__ticket_id__comments_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "ticket_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticket Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommentCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tickets/{ticket_id}/status": {
      "post": {
        "summary": "Change Status",
        "operationId": "change_status_api_tickets__ticket_id__status_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "ticket_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticket Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatusChange"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CommentCreate": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 3,
            "title": "Text"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "text"
        ],
        "title": "CommentCreate"
      },
      "Event": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "actor_name": {
            "type": "string",
            "title": "Actor Name"
          },
          "actor_role": {
            "type": "string",
            "title": "Actor Role"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "status": {
            "$ref": "#/components/schemas/Status"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "actor_name",
          "actor_role",
          "kind",
          "status",
          "text",
          "created_at"
        ],
        "title": "Event"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Profile": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "house_id": {
            "type": "string",
            "title": "House Id"
          },
          "address": {
            "type": "string",
            "title": "Address"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "role",
          "house_id",
          "address"
        ],
        "title": "Profile"
      },
      "Status": {
        "type": "string",
        "enum": [
          "new",
          "accepted",
          "in_progress",
          "resolved",
          "confirmed",
          "reopened"
        ],
        "title": "Status"
      },
      "StatusChange": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/Status"
          },
          "comment": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 3,
            "title": "Comment"
          },
          "expected_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Version"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "status",
          "comment",
          "expected_version"
        ],
        "title": "StatusChange"
      },
      "Ticket": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "house_id": {
            "type": "string",
            "title": "House Id"
          },
          "resident_id": {
            "type": "string",
            "title": "Resident Id"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "location": {
            "type": "string",
            "title": "Location"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "status": {
            "$ref": "#/components/schemas/Status"
          },
          "priority": {
            "type": "string",
            "enum": [
              "normal",
              "urgent",
              "emergency"
            ],
            "title": "Priority"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          },
          "due_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due At"
          },
          "first_response_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Response At"
          },
          "closed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Closed At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "house_id",
          "resident_id",
          "category",
          "location",
          "description",
          "status",
          "priority",
          "version",
          "created_at",
          "updated_at"
        ],
        "title": "Ticket"
      },
      "TicketAttachment": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "external_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Id"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "source",
          "type",
          "metadata",
          "created_at"
        ],
        "title": "TicketAttachment"
      },
      "TicketCreate": {
        "properties": {
          "category": {
            "type": "string",
            "enum": [
              "heating",
              "water",
              "electricity",
              "elevator",
              "cleaning",
              "yard",
              "other"
            ],
            "title": "Category",
            "default": "other"
          },
          "location": {
            "type": "string",
            "maxLength": 160,
            "minLength": 2,
            "title": "Location"
          },
          "description": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 10,
            "title": "Description"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "location",
          "description"
        ],
        "title": "TicketCreate"
      },
      "TicketDetail": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "house_id": {
            "type": "string",
            "title": "House Id"
          },
          "resident_id": {
            "type": "string",
            "title": "Resident Id"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "location": {
            "type": "string",
            "title": "Location"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "status": {
            "$ref": "#/components/schemas/Status"
          },
          "priority": {
            "type": "string",
            "enum": [
              "normal",
              "urgent",
              "emergency"
            ],
            "title": "Priority"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          },
          "due_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due At"
          },
          "first_response_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Response At"
          },
          "closed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Closed At"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/Event"
            },
            "type": "array",
            "title": "Events"
          },
          "attachments": {
            "items": {
              "$ref": "#/components/schemas/TicketAttachment"
            },
            "type": "array",
            "title": "Attachments"
          }
        },
        "type": "object",
        "required": [
          "id",
          "house_id",
          "resident_id",
          "category",
          "location",
          "description",
          "status",
          "priority",
          "version",
          "created_at",
          "updated_at",
          "events"
        ],
        "title": "TicketDetail"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}